requirements.txt
numpy>=1.24.0
opencv-python>=4.8.0
scipy>=1.11.0
matplotlib>=3.7.0
hypothesis>=6.82.0
Pillow>=10.0.0